[x86] perf/x86/mbm: Implement RMID recycling
Message-id: <1461593857-11091-15-git-send-email-jolsa@redhat.com>
Patchwork-id: 141897
O-Subject: [PATCH RHEL7.3 BZ1084618 14/16] perf/x86/mbm: Implement RMID recycling
Bugzilla: 1084618
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1084618
https://bugzilla.redhat.com/show_bug.cgi?id=1084618

upstream
========
commit 2d4de8376ff1d94a5070cfa9092c59bfdc4e693e
Author: Vikas Shivappa <vikas.shivappa@linux.intel.com>
Date: Thu Mar 10 15:32:11 2016 -0800

description
===========
RMID could be allocated or deallocated as part of RMID recycling.

When an RMID is allocated for MBM event, the MBM counter needs to be
initialized because next time we read the counter we need the previous
value to account for total bytes that went to the memory controller.

Similarly, when RMID is deallocated we need to update the ->count
variable.

Signed-off-by: Rafael Aquini <aquini@redhat.com>