[perf] Allow storage of PMU private data in event
Message-id: <1461739452-8848-2-git-send-email-jolsa@redhat.com>
Patchwork-id: 142078
O-Subject: [PATCH RHEL7.3 BZ1330700 1/2] perf: Allow storage of PMU private data in event
Bugzilla: 1330700
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1330700
https://bugzilla.redhat.com/show_bug.cgi?id=1330700

upstream
========
commit 54d751d4ad357c817907fe89db3222b97ff66db3
Author: Thomas Gleixner <tglx@linutronix.de>
Date:   Mon Feb 22 22:19:14 2016 +0000

KABI
====
The 'struct perf_event*' pointer is used within 'struct thread_struct',
making perf_event struct KABI-broken any time we change it. At the moment
we dont have any perf related function on KABI list.

description
===========
For PMUs which are not per CPU, but e.g. per package/socket, we want to be
able to store a reference to the underlying per package/socket facility in the
event at init time so we can avoid magic storage constructs in the PMU driver.

This allows us to get rid of the per CPU dance in the intel uncore and RAPL
drivers and avoids a lookup of the per package data in the perf hotpath.

Signed-off-by: Rafael Aquini <aquini@redhat.com>