[x86] mm: Drop WARN from multi-BAR check
Message-id: <1461739861-9443-1-git-send-email-jolsa@redhat.com>
Patchwork-id: 142080
O-Subject: [PATCH RHEL7.3 BZ1318419] x86/mm: Drop WARN from multi-BAR check
Bugzilla: 1318419
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1318419
https://bugzilla.redhat.com/show_bug.cgi?id=1318419

upstream
========
commit 9abb0ecdee69a2577560cc283368e490da974934
Author: Laura Abbott <labbott@fedoraproject.org>
Date:   Mon Dec 21 12:01:14 2015 -0800

description
===========
ioremapping multiple BARs produces a warning with a message "Your kernel is
fine". This message mostly serves to comfort kernel developers. Users do
not read the message, they only see the big scary warning which means
something must be horribly broken with their system. Less dramatically, the
warn also sets the taint flag which makes it difficult to differentiate
problems. If the kernel is actually fine as the warning claims it doesn't
make sense for it to be tainted. Change the WARN_ONCE to a pr_warn with the
caller of the ioremap.

brew
====
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=10935146

test
====
warning goes away on affected server

Cc: Larry Woodman <lwoodman@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>