[net] bluetooth: KABI cleanups
Message-id: <1462315287-858-22-git-send-email-dzickus@redhat.com>
Patchwork-id: 142772
O-Subject: [RHEL7 PATCH 21/22] bluetooth: KABI cleanups
Bugzilla: 1296707
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Upstream: RHEL only

net/bluetooth has 4 functions on the kabi whitelist, all of them utilizing
struct hci_dev.  That function was thrashed as part of the bluetoot rebase.

hci_alloc_dev
hci_free_dev
hci_register_dev
hci_unregister_dev

The common way to allocate this structure is to call 'hci_alloc_dev' before
passing it along to hci_register_dev().  This way allows the kernel to allocate
the memory, which means it should be safe add all the new elements of the
struct to the end.

This approach resolves most of the kabi breakage.  However, there are a few
tricky spots left.  Those can be seen through the RH_KABI_DEPRECATE calls.

Basically those 4 spots were replaced with larger structs which expands the
kabi union and trips over the Static_Assert.  I solved this by deprecating
those elements and adding them to the bottom.  Those elements are used
internally and not by the drivers, so this should be safe.

One function that was modified and broke kabi _is_ used by drivers for
registration.  I hacked around this by supporting the old style of the send
function if it is populated, otherwise utilize the new 'send' function.

Signed-off-by: Rafael Aquini <aquini@redhat.com>