[redhat] Set bluetooth and 6lowpan config options
Message-id: <1462315287-858-23-git-send-email-dzickus@redhat.com>
Patchwork-id: 142773
O-Subject: [RHEL7 PATCH 22/22] redhat: Set bluetooth and 6lowpan config options
Bugzilla: 1296707
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

Upstream: RHEL only

Enable/Disable the various config options as a result of the bluetooth
rebase.

The disabled HCIUART options are because we are missing some infrastructure
that allows them to work correctly (namely gpio which needs irq work).

The other options are straightforward.

CONFIG_6LOWPAN needs to be enabled because we enabled IEEE802154_6LOWPAN
in RHEL-7.0.  Otherwise it should be disabled because it isn't known to work
reliably yet.

Signed-off-by: Rafael Aquini <aquini@redhat.com>