[fs] nfsd: use short read as well as i_size to set eof
Message-id: <94912cb1775ce8f5ac625e4e380003989fc7ac94.1462417345.git.bcodding@redhat.com>
Patchwork-id: 142938
O-Subject: [RHEL7 PATCH fs] nfsd: use short read as well as i_size to set eof
Bugzilla: 1332694
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Use the result of a local read to determine when to set the eof flag.  This
allows us to return the location of the end of the file atomically at the
time of the read.

Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
[bfields: add some documentation]
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream-Id: ac503e4a309a3993a069750f95c2815ee5db5aa5
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1332694
Tested: cthon

Signed-off-by: Rafael Aquini <aquini@redhat.com>