[fs] nfs4: have do_vfs_lock take an inode pointer
Message-id: <713633aa1bbe6b7783bdef8c22ac800eb5cb0cdb.1462491372.git.bcodding@redhat.com>
Patchwork-id: 143063
O-Subject: [RHEL7 PATCH fs 4/5] nfs4: have do_vfs_lock take an inode pointer
Bugzilla: 1329488
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Jeff Layton <jeff.layton@primarydata.com>

Now that we have file locking helpers that can deal with an inode
instead of a filp, we can change the NFSv4 locking code to use that
instead.

This should fix the case where we have a filp that is closed while flock
or OFD locks are set on it, and the task is signaled so that it doesn't
wait for the LOCKU reply to come in before the filp is freed. At that
point we can end up with a use-after-free with the current code, which
relies on dereferencing the fl_file in the lock request.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Reviewed-by: "J. Bruce Fields" <bfields@fieldses.org>
Tested-by: "J. Bruce Fields" <bfields@fieldses.org>
Upstream-Id: 83bfff23e9ed19f37c4ef0bba84e75bd88e5cf21
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1329488
Tested: cthon
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=10976183
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>