[redhat] RHELPatchRules - add section on BZ reuse
Message-id: <20160509145721.GV20102@redhat.com>
Patchwork-id: 5302
O-Subject: [kernel team] [RHEL7 PATCH v2] redhat: RHELPatchRules - add section on BZ reuse
Bugzilla:
RH-Acked-by: Herton R. Krzesinski <herton@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

We had a conference call and email discussion with QE about requirements
and expectations on BZ reuse and we got a simple set of rules to easily
determine when reuse of BZ is allowed.

To make the process easier for engineering, QE will do the task of
classifying the BZ types at qa_ack+ time.

Also during the same discussion the procedure for handling
"full reverts" (*) was agreed and also included in this patch.

Please make sure to keep the Cc list intact as the discussion.

(*) read the definition in the patch before confusing with simple patch
reverts

v2: changes based on Rob Evers's and Herton Krzesinski's comments

Cc: aquini@redhat.com
Cc: herton@redhat.com
Cc: dzickus@redhat.com
Cc: skozina@redhat.com
Cc: prarit@redhat.com
Cc: ccui@redhat.com
Cc: revers@redhat.com
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>