[scsi] qla2xxx: Set relogin flag when we fail to queue login requests
Message-id: <1462920421-3973-41-git-send-email-cdupuis@redhat.com>
Patchwork-id: 143811
O-Subject: [PATCH RHEL 7.3 e-stor 40/41] qla2xxx: Set relogin flag when we fail to queue login requests.
Bugzilla: 1273080
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Chad Dupuis <chad.dupuis@qlogic.com>

linux commit 080c9517eaa84a2edf70382d60c46c3452869bf2

If we fail to queue an srb for an async login we should set the
relogin flag so it will be retried as the reason for the queuing
failure was most likely transient.  Failure to do this can lead to
failed paths as login is never retried if the relogin flag is not
set.

Signed-off-by: Chad Dupuis <chad.dupuis@qlogic.com>
Signed-off-by: Himanshu Madhani <himanshu.madhani@qlogic.com>
Reviewed-by: Johannes Thumshirn <jthumshirn@suse.de>
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>