[misc] cxl: Poll for outstanding IRQs when detaching a context
Message-id: <20160511185334.48211.14302.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 143884
O-Subject: [PATCH RHEL7.3 BZ1332487 2/2] cxl: Poll for outstanding IRQs when detaching a context
Bugzilla: 1332487
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1332487

Description:
------------
When detaching contexts, we may still have interrupts in the system
which are yet to be delivered to any CPU and be acked in the PSL.
This can result in a subsequent unrelated process getting an spurious
IRQ or an interrupt for a non-existent context.

This polls the PSL to ensure that the PSL is clear of IRQs for the
detached context, before removing the context from the idr.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Tested-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Acked-by: Ian Munsie <imunsie@au1.ibm.com>
Tested-by: Vaibhav Jain <vaibhav@linux.vnet.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=2bc79ffcbb817873cc43d63118008ab75181b73d

----------------
Steve Best

Signed-off-by: Rafael Aquini <aquini@redhat.com>