[include] regulator: Sync regulator/consumer.h with v4.5
Message-id: <1463503488-149391-3-git-send-email-dzickus@redhat.com>
Patchwork-id: 144596
O-Subject: [RHEL7 PATCH 002/725] regulator: Sync regulator/consumer.h with v4.5
Bugzilla: 1127975 1277866 1280133 1286932 1297039
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Upstream: RHEL-7 only

As part of the mmc rebase, calls to the regulator interface has changed
upstream.  We don't compile the REGULATOR code, but it is still expected to
have the functions correctly stubbed out.

Those calls are in include/linux/regulator/consumer.h.  Some of those fixes
include making sure the return pointers are of type ERR_PTR and not NULL.

Instead of backporting hundreds of patches that don't get compiled, just take
the simple approach and backport just the header file for now.

This works with all the arches and compiles, boots and tests fine.

Signed-off-by: Rafael Aquini <aquini@redhat.com>