[x86] perf/x86/intel/rapl: Convert it to a per package facility
Message-id: <1463569977-29479-9-git-send-email-jolsa@redhat.com>
Patchwork-id: 145683
O-Subject: [PATCH RHEL7.3 BZ1336447 008/324] perf/x86/intel/rapl: Convert it to a per package facility
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit 9de8d686955b0e8e27847ed4edbbcd280f3fd853
Author: Thomas Gleixner <tglx@linutronix.de>
Date: Mon Feb 22 22:19:26 2016 +0000

description
===========
RAPL is a per package facility and we already have a mechanism for a dedicated
per package reader. So there is no point to have multiple CPUs doing the
same. The current implementation actually starts two timers on two CPUs if one
does:

	perf stat -C1,2 -e -e power/energy-pkg ....

which makes the whole concept of 1 reader per package moot.

What's worse is that the above returns the double of the actual energy
consumption, but that's a different problem to address and cannot be solved by
removing the pointless per cpuness of that mechanism.

Signed-off-by: Rafael Aquini <aquini@redhat.com>