[x86] perf/x86/intel/pt: Generate PMI in the STOP region as well
Message-id: <1463569977-29479-26-git-send-email-jolsa@redhat.com>
Patchwork-id: 145605
O-Subject: [PATCH RHEL7.3 BZ1336447 025/324] perf/x86/intel/pt: Generate PMI in the STOP region as well
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit ab92b232ae05c382c3df0e3d6a5c6d16b639ac8c
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Tue May 10 16:18:32 2016 +0300

description
===========
Currently, the PT driver always sets the PMI bit one region (page) before
the STOP region so that we can wake up the consumer before we run out of
room in the buffer and have to disable the event. However, we also need
an interrupt in the last output region, so that we actually get to disable
the event (if no more room from new data is available at that point),
otherwise hardware just quietly refuses to start, but the event is
scheduled in and we end up losing trace data till the event gets removed.

For a cpu-wide event it is even worse since there may not be any
re-scheduling at all and no chance for the ring buffer code to notice
that its buffer is filled up and the event needs to be disabled (so that
the consumer can re-enable it when it finishes reading the data out). In
other words, all the trace data will be lost after the buffer gets filled
up.

This patch makes PT also generate a PMI when the last output region is
full.

Signed-off-by: Rafael Aquini <aquini@redhat.com>