[tools] perf hists browser: Skip scripting when perf.data file not available
Message-id: <1463569977-29479-50-git-send-email-jolsa@redhat.com>
Patchwork-id: 145666
O-Subject: [PATCH RHEL7.3 BZ1336447 049/324] perf hists browser: Skip scripting when perf.data file not available
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit b1baae89197e21cd115e9493d5a17f18fca81e6a
Author: Namhyung Kim <namhyung@kernel.org>
Date: Tue Jan 26 15:37:30 2016 -0300

description
===========
The script and data-switch context menu are only meaningful when it
deals with a data file.  So add a check so that it cannot be shown when
perf-top is run.

[ Use goto skip_scripting instead of two is_report_browser() tests ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>