[tools] perf hists: Support filtering in hierarchy mode
Message-id: <1463569977-29479-200-git-send-email-jolsa@redhat.com>
Patchwork-id: 145812
O-Subject: [PATCH RHEL7.3 BZ1336447 199/324] perf hists: Support filtering in hierarchy mode
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit 155e9afff77916931f615a394cef187b342530dc
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Feb 25 00:13:38 2016 +0900

description
===========
The hists__filter_hierarchy() function implements filtering in hierarchy
mode.  Now we have hist_entry__filter() so use it for entries in the
hierarchy.  It returns 3 kind of values.

A negative value means that it's not filtered by this type.  It marks
current entry as filtered tentatively so if a lower level entry removes
the filter it also removes the all parent so that we can find the entry
in the output.

Zero means it's filtered out by this type. A positive value means it's
not filtered so it removes the filter and shows in the output.  In these
cases, it moves to next entry since lower level entry won't match by
this type of filter anymore.  Thus all children will be filtered or not
together.

Signed-off-by: Rafael Aquini <aquini@redhat.com>