[tools] perf hists: Add more helper functions for the hierarchy mode
Message-id: <1463569977-29479-218-git-send-email-jolsa@redhat.com>
Patchwork-id: 145879
O-Subject: [PATCH RHEL7.3 BZ1336447 217/324] perf hists: Add more helper functions for the hierarchy mode
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit a7b5895b91fb97f2b0dcc2e3ce47413c18d19ca5
Author: Namhyung Kim <namhyung@kernel.org>
Date: Fri Feb 26 21:13:16 2016 +0900

description
===========
The hists__overhead_width() is to calculate width occupied by the
overhead (and others) columns before the sort columns.

The hist_entry__has_hiearchy_children() is to check whether an entry has
lower entries (children) in the hierarchy to be shown in the output.
This means the children should not be filtered out and above the percent
limit.

These two functions will be used to show information when all children
of an entry is omitted by the percent limit (or filter).

Signed-off-by: Rafael Aquini <aquini@redhat.com>