[tools] perf tools: Explicitly declare inc_group_count as a void function
Message-id: <1463569977-29479-251-git-send-email-jolsa@redhat.com>
Patchwork-id: 145746
O-Subject: [PATCH RHEL7.3 BZ1336447 250/324] perf tools: Explicitly declare inc_group_count as a void function
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit 07ef7574458369cb0345facc748e964af68a75f4
Author: Colin Ian King <colin.king@canonical.com>
Date: Mon Mar 7 16:44:37 2016 -0300

description
===========
The return type is not defined, so it defaults to int, however, the
function is not returning anything, so this is clearly not correct. Make
it a void function.

Signed-off-by: Rafael Aquini <aquini@redhat.com>