[tools] perf inject: Hit all DSOs for AUX data in JIT and other cases
Message-id: <1463569977-29479-252-git-send-email-jolsa@redhat.com>
Patchwork-id: 145831
O-Subject: [PATCH RHEL7.3 BZ1336447 251/324] perf inject: Hit all DSOs for AUX data in JIT and other cases
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit 640dad47988ec4b734d71934be103bb6e931279f
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Mon Mar 7 16:44:38 2016 -0300

description
===========
Currently, when injecting build ids, if there is AUX data then 'perf
inject' hits all DSOs because it is not known which DSOs the trace data
would hit.

That needs to be done for JIT injection also, and in fact there is no
reason to distinguish what kind of injection is being done.  That is,
any time there is AUX data and the HEADER_BUID_ID feature flag is set,
and the AUX data is not being processed, then hit all DSOs.  This patch
does that.

Signed-off-by: Rafael Aquini <aquini@redhat.com>