[tools] perf tools: Remove nr_sort_keys field
Message-id: <1463569977-29479-277-git-send-email-jolsa@redhat.com>
Patchwork-id: 145775
O-Subject: [PATCH RHEL7.3 BZ1336447 276/324] perf tools: Remove nr_sort_keys field
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit 86e3ee5224c17b7967aac39aa15539393c144de7
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Mar 9 22:47:01 2016 +0900

description
===========
The nr_sort_keys field is to carry the number of sort entries in a
hpp_list or hists to determine the depth of indentation of a hist entry.
As it's only used in hierarchy mode and now we have used nr_hpp_node for
this reason, there's no need to keep it anymore.  Let's get rid of it.

Signed-off-by: Rafael Aquini <aquini@redhat.com>