[tools] perf tools: Recalc total periods using top-level entries in hierarchy
Message-id: <1463569977-29479-278-git-send-email-jolsa@redhat.com>
Patchwork-id: 145862
O-Subject: [PATCH RHEL7.3 BZ1336447 277/324] perf tools: Recalc total periods using top-level entries in hierarchy
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit f7fb538afea55383a9383dac5c56887c601af5f4
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Mar 9 22:47:02 2016 +0900

description
===========
When hierarchy mode is enabled, each entry in a hierarchy level shares
the period.  IOW an upper level entry's period is the sum of lower level
entries.  Thus perf uses only one of them to calculate the total period
of hists.  It was lowest-level (leaf) entries but it has a problem when
it comes to filters.

If a filter is applied, entries in the same level will be filtered or
not.  But upper level entries still have period of their sum including
filtered one.  So total sum of upper level entries will not be same as
sum of lower level entries.

This resulted in entries having more than 100% of overhead and it can be
produced using perf top with filter(s).

Signed-off-by: Rafael Aquini <aquini@redhat.com>