[tools] tools lib traceevent: Do not reassign parg after collapse_tree()
Message-id: <1463569977-29479-322-git-send-email-jolsa@redhat.com>
Patchwork-id: 145781
O-Subject: [PATCH RHEL7.3 BZ1336447 321/324] tools lib traceevent: Do not reassign parg after collapse_tree()
Bugzilla: 1336447
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1336447
https://bugzilla.redhat.com/show_bug.cgi?id=1336447

upstream
========
commit 106b816cb46ebd87408b4ed99a2e16203114daa6
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Wed May 11 15:09:36 2016 -0400

description
===========
At the end of process_filter(), collapse_tree() was changed to update
the parg parameter, but the reassignment after the call wasn't removed.

What happens is that the "current_op" gets modified and freed and parg
is assigned to the new allocated argument. But after the call to
collapse_tree(), parg is assigned again to the just freed "current_op",
and this causes the tool to crash.

The current_op variable must also be assigned to NULL in case of error,
otherwise it will cause it to be free()ed twice.

Fixes: 42d6194d133c ("tools lib traceevent: Refactor process_filter()")
Signed-off-by: Rafael Aquini <aquini@redhat.com>