[netdrv] i40e/i40evf: fix unicast mac address add
Message-id: <1463589233-31199-15-git-send-email-sassmann@redhat.com>
Patchwork-id: 146101
O-Subject: [RHEL7.3 PATCH 014/320] i40e/i40evf: fix unicast mac address add
Bugzilla: 1269238 1272322 1274176 1281279 1293277 1310436
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ken Cox <jkc@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Shannon Nelson <shannon.nelson@intel.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1274176
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11047211

When using something like "ip maddr add ..." to add another unicast mac
address to the netdev, the mac address comes into the set_rx_mode handler
in the multicast list whether it is a unicast or multicast address.
This was confusing the code when it was trying to search for addresses
that needed to be deleted from the VSI, because it was looking for the
VSI unicast address in the netdev unicast list.  The result was that a
new unicast address would get added to the VSI list and then immediately
removed, and would never actually make it down into the hardware.

This patch removes the separation from unicast and multicast in the search
for filters to be deleted.  It also simplifies the logic a little with a
jump to the bottom of the loop when an address is found.  Now it doesn't
matter which netdev list the address is hiding in, we'll check them all.

Change-ID: Ie3685a92427ae7d2212bf948919ce295bc7a874c
Signed-off-by: Shannon Nelson <shannon.nelson@intel.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit 2f41f3358672dfda67c1e254f1e823d98e6a3099)
Signed-off-by: Rafael Aquini <aquini@redhat.com>