[netdrv] i40e/i40evf: Drop outer checksum offload that was not requested
Message-id: <1463589233-31199-222-git-send-email-sassmann@redhat.com>
Patchwork-id: 146333
O-Subject: [RHEL7.3 PATCH 221/320] i40e/i40evf: Drop outer checksum offload that was not requested
Bugzilla: 1269238 1272322 1274176 1281279 1293277 1310436
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ken Cox <jkc@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Alexander Duyck <aduyck@mirantis.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1274176
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11047211

The i40e and i40evf drivers contained code for inserting an outer checksum
on UDP tunnels.  The issue however is that the upper levels of the stack
never requested such an offload and it results in possible errors.

In addition the same logic was being applied to the Rx side where it was
attempting to validate the outer checksum, but the logic there was
incorrect in that it was testing for the resultant sum to be equal to the
header checksum instead of being equal to 0.

Since this code is so massively flawed, and doing things that we didn't ask
for it to do I am just dropping it, and will bring it back later to use as
an offload for SKB_GSO_UDP_TUNNEL_CSUM which can make use of such a
feature.

As far as the Rx feature I am dropping it completely since it would need to
be massively expanded and applied to IPv4 and IPv6 checksums for all parts,
not just the one that supports Tx checksum offload for the outer.

Signed-off-by: Alexander Duyck <aduyck@mirantis.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit a9c9a81f5892eb984234223399ee624f7dbd15e8)
Signed-off-by: Rafael Aquini <aquini@redhat.com>