[netdrv] i40e/i40evf: Factor out L4 header and checksum from L3 bits in TSO path
Message-id: <1463589233-31199-224-git-send-email-sassmann@redhat.com>
Patchwork-id: 146306
O-Subject: [RHEL7.3 PATCH 223/320] i40e/i40evf: Factor out L4 header and checksum from L3 bits in TSO path
Bugzilla: 1269238 1272322 1274176 1281279 1293277 1310436
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ken Cox <jkc@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Alexander Duyck <aduyck@mirantis.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1274176
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=11047211

This patch makes it so that the L4 header offsets and such can be ignored
when dealing with the L3 checksum and length update.  This is done making
use of two things.

First we can just use the offset from the L4 header to the start of the
packet to determine the L4 offset, and from that we can then make use of
the data offset to determine the full length of the headers.

As far as adjusting the checksum to remove the length we can simply add the
inverse of the length instead of having to recompute the entire
pseudo-header without the length.  In the case of an IPv6 header this
should be significantly cheaper since we can make use of a value we already
needed instead of having to read the source and destination address out of
the packet.

Signed-off-by: Alexander Duyck <aduyck@mirantis.com>
Tested-by: Andrew Bowers <andrewx.bowers@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit c49a7bc3308099a8d5f9e2e38adfc5ab969804aa)
Signed-off-by: Rafael Aquini <aquini@redhat.com>