[fs] revert "libxfs: pack the agfl header structure so XFS_AGFL_SIZE is correct"
Message-id: <ac2b88ba-6d63-b39c-a3ee-4c77f53c65bc@redhat.com>
Patchwork-id: 146502
O-Subject: [PATCH RHEL7] [fs] revert "libxfs: pack the agfl header structure so XFS_AGFL_SIZE is correct"
Bugzilla: 1336918
RH-Acked-by: Brian Foster <bfoster@redhat.com>
RH-Acked-by: Bill O'Donnell <billodo@redhat.com>

Bugzilla: 1336918

The commit in question was intended to resolve an incompatibility
between 32-bit and 64-bit formatted filesystems; however, it also
inadvertently caused issues if a filesystem was formatted prior
to the change, and grown after the change.  Because in RHEL7 we
only ship 64-bit kernel and xfsprogs, it is safest to simply
revert this change in RHEL7.

Upstream will have a more complex solution.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>