[redhat] spec: conflict with older xfsprogs
Message-id: <d4d33408-8135-645a-b652-ce855220c5cb@redhat.com>
Patchwork-id: 146503
O-Subject: [PATCH RHEL7] [fs] conflict with older xfsprogs
Bugzilla: 1333088
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Brian Foster <bfoster@redhat.com>

Bugzilla: 1333088

The commit:

 1c6508a0 [fs] xfs: validate metadata LSNs against log on v5 superblocks

results in new tests which may flag the filesystem as corrupt,
and suggest to the user that they should run xfs_repair >= v4.3

Add a new Conflicts: to the kernel specifile, to ensure that
version is installed along with the RHEL7.3 kernel.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
---

p.s. I'd be happy to have people who think hard about kernel packaging
confirm that this Conflicts is the right approach, and a good idea ...

Signed-off-by: Rafael Aquini <aquini@redhat.com>