[sound] alsa: hda - Fix white noise on Dell M3800
Message-id: <146374421683-104-patchset-jkysela@redhat.com>
Patchwork-id: 146635
O-Subject: [RHEL 7.3 PATCH 103/485] ALSA: hda - Fix white noise on Dell M3800
Bugzilla: 1288993
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

The M3800 is very minor workstation variant of the XPS 15 which has
already been patched for this issue.  I figured it's probably more
important for this version of the laptop to be patched than the
regular XPS as Dell sells is pre-configured with Ubuntu to be used as
a Linux workstation.  I have tested the patch on my the hardware on
Linux 4.2.0.

Signed-off-by: Niranjan Sivakumar <ns253@cornell.edu>
Cc: <stable@vger.kernel.org> # v4.1+
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 467e1436ba85f78b8c4610c4549eb255a8211c42)
rhbz#1288993
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>