[sound] alsa: usb-audio: harmless underflow in snd_audigy2nx_led_put()
Message-id: <146374421683-112-patchset-jkysela@redhat.com>
Patchwork-id: 146579
O-Subject: [RHEL 7.3 PATCH 111/485] ALSA: usb-audio: harmless underflow in snd_audigy2nx_led_put()
Bugzilla: 1288993
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

We want to verify that "value" is either zero or one, so we test if it
is greater than one.  Unfortunately, this is a signed int so it could
also be negative.  I think this is harmless but it introduces a static
checker warning.  Let's make "value" unsigned.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: e87359efcaf45efc41dbaa6ec25eb26705300cee)
rhbz#1288993
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>